fix(KIInventory): add changes from SonarQube
